feat[next][dace]: GTIR-to-DaCe lowering of map-reduce (only full connectivity) (#1683)
This PR adds support for lowering of `map_` and `make_const_list`
builtin functions. However, the current implementation only supports
neighbor tables with full connectivity (no skip values). The support for
skip values will be added in next PR.

To be noted:
- This PR generalizes the handling of tasklets without arguments inside
a map scope. The return type for `input_connections` is extended to
contain a `TaskletConnection` variant, which is lowered to an empty edge
from map entry node to the tasklet node.
- The result of `make_const_list` is a scalar value to be broadcasted on
a local field. However, in order to keep the lowering simple, this value
is represented as a 1D 1-element array (`shape=(1,)`).
# Conflicts:
#	src/gt4py/next/program_processors/runners/dace/gtir_python_codegen.py
#	src/gt4py/next/program_processors/runners/dace_common/utility.py