feat[next][dace]: Stop carrying reduce identity in args context (#1704)
This PR implements an alternative design to local expressions on lists
containing skip values. The previous design was based on the assumption
that such operations only exist in the context of a reduction (which is
the most common case, but necessarily the only one) and was carrying the
reduce identity value from a reduce node to the visiting context of the
arguments and all their child nodes. The reduce identity value was used
till fill the skip values when building lists of neighbors or when
applying map expressions. Although quite effective for reduce
expressions, this design led to a very complicated code. Besides, this
design was mixing lowering with optimization, by implicitly optimizing
the SDFG.

With this PR, we stop carrying the reduce identity to the arguments
context. Instead, in presence of skip values we just write a dummy
value. When it is time to reduce, we override the dummy value with the
reduce identity. In this way, the reduce identity is only used in the
context of the reduction expression.