feat[next]: enable gtir.embedded (and add support for Lists in output) (#1703)
Lists returned from an `as_fieldop`ed stencil will be turned into a local dimension of the the resulting field.
In case of a `make_const_list`, a magic local dimension `_CONST_DIM` is used. This is a hack, but localized to `itir.embedded`. A clean implementation will probably involve to tag the `make_const_list` with the neighborhood it is meant to be used with.
# Conflicts:
#	src/gt4py/next/iterator/embedded.py
#	src/gt4py/next/iterator/type_system/type_specifications.py
#	tests/next_tests/unit_tests/iterator_tests/test_embedded_field_with_list.py