feat[next]: Add memory and disk-based caching to more workflow steps (#1690)
Add memory and disk-based caching to other workflow steps and,
therefore, removing unnecessary overhead of Program calls and
significantly improving time to first computed value.

Changes:

- setting `cached = True` for `func_to_past_factory`
- wrapping the GTFN code generation into a `CachedStep` (using
`Diskcache`) which is activated when setting
`otf_workflow__cached_translation=True`, similar as in
[PR#1474](https://github.com/GridTools/gt4py/pull/1474) (without
CachedStep)
- Fixing hash function of `ProgramDefinition`
- New tests for added functionality

This leads to a runtime decrease of about 25% for PMAP-G in the
advect-uniform testcase (5 hours) after caches are populated.

TODOs:
- [x] improving hash functions of `fingerprint_stage`

---------

Co-authored-by: Till Ehrengruber <till.ehrengruber@cscs.ch>
Co-authored-by: Enrique G. Paredes <enriqueg@cscs.ch>
# Conflicts:
#	tests/next_tests/unit_tests/program_processor_tests/codegens_tests/gtfn_tests/test_gtfn_module.py