feat[next][dace]: Use offset_type to represent neighborhood information for local dimensions (#1734)
This PR adopts the `offset_type` design concept implemented in #1703 for
Embedded-GTIR and applies it to the DaCe-GTIR backend.

The only functional change is that the if-builtin is now expected to
return the exact same data type, including the same `offset_type` if a
local dimension is present in the result field. This change required
updates to `test_gtir_reduce_with_cond_neighbors`.