feat[next]: remove NeighborTableOffsetProvider, use gtx.as_connectivity (#1729)
User-facing change: use `gtx.as_connectivity` to create a
connectivity/neighbor table instead of `NeighborTableOffsetProvider`
which is deprecated (and the backward-compatible mechanism broken for
some use-cases).

The internal concepts of `Connectivity` and `NeighborTable` are updated.
`ConnectivityType` is introduced which contains the compile-time info of
a `Connectivity`. See ADR 19.
Additionally, the compile-time info is used (instead of the run-time
connectivities) in many places of the toolchain when possible.
# Conflicts:
#	src/gt4py/next/program_processors/runners/dace_iterator/itir_to_tasklet.py
#	src/gt4py/next/type_system/type_specifications.py
#	tests/next_tests/integration_tests/multi_feature_tests/ffront_tests/test_ffront_fvm_nabla.py
#	tests/next_tests/unit_tests/iterator_tests/test_runtime_domain.py