feat[dace][next]: Fixing strides in optimization (#1782)
Added functionality to properly handle changes of strides.

During the implementation of the scan we found that the strides were not handled properly.
Most importantly a change on one level was not propagated into the next levels, i.e. they were still using the old strides.
This PR Solves most of the problems, but there are still some issues that are unsolved:
- Views are not adjusted yet (Fixed in [PR@1784](https://github.com/GridTools/gt4py/pull/1784)).
- It is not properly checked if the symbols of the propagated strides are safe to introduce into the nested SDFG.

The initial functionality of this PR was done by Edoardo Paone (@edopao).

---------
Co-authored-by: edopao <edoardo.paone@cscs.ch>
# Conflicts:
#	tests/next_tests/unit_tests/program_processor_tests/runners_tests/dace_tests/transformation_tests/test_strides.py