fix[dace]: Fixed a bug in `gt_make_transients_persistent()` (#1831)
Make a transient persistent essentially makes it a global, it is thus
shared among different iterations of a Map.
Consider the following:
```c++
#pragma omp parallel for
for(int k = 0; k != N; ++k)
{
        double b = foo(k);
        bar(b);
};
```
In the above code each iteration has its own local copy of `b` that is
not shared with other threads.
If `b` would be declared persistent, the code essentially becomes
```c++
double* b = new double;
#pragma omp parallel for
for(int k = 0; k != N; ++k)
{
        *b = foo(k);
        bar(*b);
};
```
i.e. now `b` is shared among the different threads and we have a data
race.

To avoid this situation we have to ensure that there is no data race and
we do this by requiring that a data descriptor that should be turned
into a persistent one, can not have an AccessNode inside a scope other
than the top scope.

Note that this restriction is stronger than necessary and could be
relaxed, but it might be very difficult to figuring out if in a
particular case it is possible or not.