fix[next][dace]: handle if-expression as non-exclusive in field view (#1824)
The implementation of scan field operator introduced a bug in the
lowering of if-statements to SDFG. The scan requires extended lowering
support for iterator view, therefore #1790 introduced the lowering of
local-if with exclusive branch execution. However, attention was not
paid to enable the exclusive behavior of if-expressions only in iterator
view, that is only in the scope of scan field operators. Regular field
operators should instead lower if-expressions to a tasklet, because the
field view behavior is that of a local select.