test[next]: enable gpu test for 1d scan on dace backend (#1854)
The dace backend did not support the case of scan on a 1D vertical
array, when lowered to GPU code. After upgrading the dace package to
latest main and adopting the `LoopRegion` construct for the lowering of
scan, this case is no longer an issue. This PR just re-enables the scan
test case.