fix[dace][next]: Updated `gt_auto_optimize()` (#1889)
The problem was that in the auto optimizer the "set stride correctly"
function was called after the "go to GPU transformation". The GPU
transformation will call `CopyToMap` transformation, this is needed such
that we can set the GPU block size and their order. However, the
decision if we call `CopyToMap` also depends on the strides, so we need
to set them first.

---------

Co-authored-by: edopao <edoardo16@gmail.com>