fix[next][dace]: Update cuda codegen for concat_where (#1906)
We use `CopyToMap` in CUDA lowering for copies between arrays that do
not necessarily have the same strides. This happens in case of
`concat_where`, where we copy a source array into a subset of the
destination array.

For this reason, the option `ignore_strides` must be set to `True`
(`False` by default).