fix[dace][next]: Fixed some undefined behaviour in the chain remover. (#1910)
The issue was discovered by Edoardo (@edopao).
The underlying problem is that `get_src_subset()` tries to set the
direction of the Memlet, for that it looks for the source of the data.
However, when we called it then the source is unspecific because the new
edge is already there and the old one was not removed. This might
implicitly change the direction of the Memlet.

I have removed the calls where they are potentially dangerous and added
some counter measure to avoid the problem.