feat[dace][next]: Updated GPU Transformation Scheme (#1904)
This PR slightly changes how the GPU transformation works.
It mainly changes how trivial Maps are eliminated.
First, there is now a dedicated function, `gt_remove_trivial_gpu_maps()`
for this. Furthermore, before it was only using the `TrivialGPUMapElimination`
transformation, which tries to promote and fuse the trivial maps, but
only with downstream maps. Now there is a second stage that tries to
fuse the trivial maps together. This is mostly to reduce the number of
kernel calls that we are doing.