ci: read python versions from files (#1919)
Encode in files the python versions used for testing, linting and other
CI and development tasks.

- `.default-python-version` contains the python version used by default
for development tasks (previously `.python-version` was used, but it
conflicted with other tools like pyenv).
- `.python-versions` contains all supported python versions. This file
is read by `uv` to set up the development environment. To avoid
conflicts, the first version in the file should match the content of
`default-python-version`. This file is also used by a new github action
which is used by all CI jobs where the python version is a factor of the
strategy matrix.

Additionally, the nox task running the example notebooks is not longer
executed in parallel. This change is unrelated to this PR, but it seems
necessary to have stable results on CI.