feat[next]: Allow using compile time connectivity information only (#1862)
The domain inference relies on either runtime connectivities or compile
time information on the size of the domains (`symbolic_domain_sizes`),
whereas the later is used in icon4py, in particular the blue line, where
runtime information is not available. The approach taken in this PR is
to allow passing either an `common.OffsetProvider` or an
`common.OffsetProviderType`.