feat[next][dace]: setup cached and non-cached DaCe backends (#1941)
The `run_dace_cpu` and `run_dace_gpu` backends were using the gt4py
cache. Following the GTFN pattern, this PR introduces separate backends
(`run_dace_cpu_cached` and `run_dace_gpu_cached`) to enable the cache.
The former backends are configured without cache, and are used in unit
tests.
At the same time, this PR improves the cache configuration, by adding
caching of the translation stage. The former cache configuration was
only caching the compiled programs. With this fix, it is possible to
load translated programs (SDFGs) from disk cache.

Note that this PR does not address the limitation on test parallelism
(slow down observed when we increase the number of pytest workers).