fix[dace][next]: Added a second way to specify the leading dimension. (#1952)
Before it was only possible to specify which dimension should be
considered as "leading dimensions", i.e. associated to the stride 1.
This PR adds a second method, which can be selected through the
`leading_kind` argument, which is a GT4Py `DimensionKind` enum value.
This allows to generally specify if horizontal or vertical dimensions
are leading dimensions.

---------

Co-authored-by: edopao <edoardo16@gmail.com>