feat[next]: Asynchronous pre-compilation (#1951)
Adds `Program.compile()` which pre-compiles the program. Requires the
offset_provider_type to be passed as argument to `compile` or
connectivities added to the program with `with_connectivities()`.

Additionally, the compilation will be placed in a ThreadPoolExecutor to
allow async compilation.

Note, in a multi-rank MPI setup there is no mechanism that distributes
compilation. The first rank that reaches the `compile()` will compile,
the others will spend their thread in waiting for getting the lock, then
eventually getting the lock and loading the compiled program. The next
threads of each MPI rank will get to the next program. Again, it could
be the same rank that locks first.