feat[dace][next]: Updated `LoopBlocking` (#1973)
This PR updates the classification function of the `LoopBlocking`
transformation. The transformation is now able to detect is a
nested Map depends on the blocking parameter or not, thus
the map can be moved outside of the inner map.
Furthermore, the classification function is now also able to handle
reduction nodes.

---------

Co-authored-by: Hannes Vogt <hannes@havogt.de>