fext[next]: gtfn scan projector (#1987)
Hides a projector from the temporary extraction and applies it as
projector on a scan operator in gtfn.

A projector is a make_tuple/tuple_get operation (possibly nested) on the
same expression.

This enables cases where the frontend doesn't return all output of a
scan and avoids creating temporaries for unused tuple elements.