fix[dace][next]: Fixed `MapFusion`'s `relocate_node()` Method (#2009)
Fixes the behaviour of MapFusion in cases where the Map is nested.
It turned out that the consolidation of edges was not fully working,
this was why the error did not materialize yet.
Essentially, this PR introduces the same changes as [DaCe
PR#1993](https://github.com/spcl/dace/pull/1993) does.

---------

Co-authored-by: edopao <edoardo16@gmail.com>