fix[dace][next]: Fixed `MoveDataflowIntoIfBody` (#2002)
This PR fixes how the if mover handling the "patching", i.e. how it
makes the data that are needed inside available. The previous
implementation did not considered the case when an external AccessNode
had multiple edges to the enclosing Map. It detected correctly that it
only had to create it once, however the clean up was wrong, as it would
remove the inner edge but leave the outer edge, the one connecting the
AccessNode with the enclosing Map alive.
This created dangling connectors.