feat[next][dace]: Improve SDFG fast call (#2029)
This PR reduces the overhead of the decorator function for program call
in the dace backend. It achieves this by implementing the
`bindings`stage in the backend workflow, which dynamically generates
Python code for mapping the call arguments in the gt4py canonical form
to the SDFG arglist canonical form. Since the order and the data types
of the arguments is known at compile time, the generated code is
tailored to each specific program and optimized for minimum overhead.
This code is then loaded as a Python module in the compilation stage and
executed in the program call decorator to update the values of the SDFG
argument list.

Refer to the unit test for an example of generated binding code.

Core changes:
- Created a new `bindings` stage in the dace backend.
- Added `make_persistent` option to translation stage and bindings stage
in dace backend, that allows to _optimize_ code generation for
applications like ICON4Py where the field layout does not change across
multiple program calls. This allows to omit the code to update the SDFG
symbols for field size and strides.

Additional change:
- Rename connectivity fields `connectivity_` to `gt_conn_`