refactor[next]: refactor embedded execution context utilities (#2046)
Refactor the embedded execution context handling with a simpler
implementation suitable more suitable for non-async code and allowing
the use of other context vars outside of this module. Getter functions
for the context vars are now in place to force the use of the context
manager as the only safe way to modify the contextvars. New unit tests
have been also added.

Additionally, the test utilities in `cases` contain a couple of fixtures
for cases without backend, which involved a slight refactoring of the
module. This new fixture is used in an unrelated gtfn cache test which
triggered a hidden bug in the embedded execution context.