feat[next]: Performance timers (#1978)
Adds a config option `COLLECT_METRICS_LEVEL` to collect performance
metrics with different levels of granularity. If enabled it will collect
for each program some of the following metrics:
- total runtime: if programs are not pre-compiled this contains
everything from call to result.
- compute time: time spend on the C++ side in gtfn, for CUDA this adds a
`deviceSynchronize` after the kernel which may impact total runtime of
an application.

Additional timers of the transformation/compilation pipeline could be
added like the following:
- create a new `TimedStep` that wraps any workflow step
- if metrics collection is enabled, we could wrap the steps in
`OTFCompileWorkflow` with the `TimedStep`.

---------

Co-authored-by: Enrique Gonzalez Paredes <enriqueg@cscs.ch>