feat[dace][next]: Updated Edge Consolidation In `MapFusion` (#2061)
During top level data flow optimization we had to suppress the
consolidation of edges to avoid an extension. A behaviour introduced by
[PR#1992](https://github.com/GridTools/gt4py/pull/1992). The example is
that a Map reads from `tmp[__i, 0]`, where `__i` is the Map parameter,
is fused with a Map that reads `tmp[__i, 80]`. If the edges are
consolidated then we will end up with one edge, whose subset is
`tmp[__i, 0:80]`, whcih might causes trouble. Note the more severe
problem was the `ConsolidateEdge` pass in simplify.

However, it turned out that this might limit MapFusion, thus a new flag
was added to allow consolidation if it does not lead to an extension.
In the example above no consolidation will be done, i.e. we will still
end up with two edges. However, if the first Map reads from `tmp[__i,
4:77]` and the second from `tmp[__i, 6:50]` then consolidation is
performed.

Note:
The MapFusion that ships with GT4Py will be removed by the one that
ships with DaCe once [PR#1987](https://github.com/spcl/dace/pull/1987)
has been merged. The changes in this PR will be backported to DaCe where
also the unit tests will be implemented.

---------

Co-authored-by: Ioannis Magkanaris <iomagkanaris@gmail.com>