feat[next]: GTIR concat_where frontend (remove chained comparison) (#1998)
GTIR concat_where frontend parts extracted from #1713. Credits to @SF-N
and @tehrengruber .

This PR removes the support for chained comparisons like `field_a <
field_b < field_c` which cannot be supported in embedded (same as `0 <
Dim < 42`) because it translates to a boolean scalar comparison as it is
evaluated as `(field_a < field_b) and (field_b < field_c)` with `and`
instead of `&`.

---------

Co-authored-by: tehrengruber <till.ehrengruber@cscs.ch>