feat[dace][next]: Added the "Splitting Tools" (#2058)
This is a set of tools that allows to perform splits of edges and
AccessNodes.
They should more be seen as a tool box.

In addition this PR also updates the `SplitAccessNode` transformation,
it now also uses the tool box and adds the `SplitConsumerMemlet` that
allows to split consumer edges, such that `SplitAccessNode is more
applicable.


---------

Co-authored-by: Ioannis Magkanaris <ioannis.magkanaris@cscs.ch>