fix[dace][next]: Allowed Horizontal Promotion in `gt_auto_optimize()` (#2063)
In [PR#1992](https://github.com/GridTools/gt4py/pull/1992) the behaviour
of the consolidation of edges during the top level data flow
optimization was changed and this caused a series of performance
regressions.
One of them was located in `19_to_20` velocity advection stencil in
ICON4Py, where a horizontal promotion was previously present, but not
any longer, which caused a reduction in performance.
In the end it turned out that we forget to merge a [PR in ICON4Py in
which this very optimization was done
manually](https://github.com/C2SM/icon4py/pull/735) thus we decided to
enable it by default.

The PR also adds several unit tests to fully ensure the correct
behaviour.

---------

Co-authored-by: Ioannis Magkanaris <ioannis.magkanaris@cscs.ch>