fix[next][dace]: Do not apply map_promoter on symbolic range (#2071)
When a 1d temporary is consumed by different 2d maps, with symbolic
range, we can promote the 1d producer map to a 2d map only if the range
of the new dimension is known to be not empty. The issue found in
icon4py (stencil `39_to_40`, without constant substitution) was that the
temporary was no longer produced, after applying MapPromotion, because
the new dimension was empty. The other consumer maps, that actually
needed the temporary data, found uninitialized data (it happened to be
zeros).