refactor[cartesian]: fix mypy issues for pyext_builder (#2080)
## Description

`mypy` was configured to ignore all errors in `pyext_builder`. This was
a temporary workaround (I guess from when `mypy` was first added). It
was easy enough to "fix" what `mypy` was complaining about:

1. The overload that would "tells mypy to accept unpacking kwargs"
wasn't really working. It also seems like the code isn't used, so I'm
getting rid of it.
2. `mypy` can't know that `setuptools_args["script_args"]` is a list
because `setuptools_args` is a dictionary that does also contain values
that aren't lists. This is fine, we know, so we can just ignore the
error with a line override.

## Requirements

- [x] All fixes and/or new features come with corresponding tests.
  All good as long as the CI is still green.
- [ ] Important design decisions have been documented in the appropriate
ADR inside the [docs/development/ADRs/](docs/development/ADRs/README.md)
folder.
  N/A

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>