chore: use precommit local hooks with uv and update dev scripts (#2086)
Update pre-commit configuration to use local hooks for python tools
tracked by `uv` to keep versions in sync without duplication of the
version information.

Additionally, put in place a new developer-friendly infrastructure to
run custom development scripts in a simple and documented way. For now,
only a script to update dependencies has been added. The main entry
point for the scripts is the `./dev-scripts.py` file. If run without
options (or with the `--help` option) displays the usage help for all
commands, and the same applies for the subcommands. For example:

```bash
./dev-scripts.py --help    # Show list of commands
./dev-scripts.py update --help    # Show usage for the 'update' command (including list of subcommands)
./dev-scripts.py update pre-commit --help    # Show usage help for the pre-commit 'subcommand' of the 'update' command
```

Finally, update the versions of tools in the pre-commit hooks which
affects the formatting of several source code files.