fix[dace][next]: Get correct symbol mapping in MoveDataflowIntoIfBody (#2154)
The `MoveDataflowIntoIfBody` transformation moves part of a dataflow
into the nested SDFG of an If-Block. Some symbols need to be mapped to
symbols in the parent SDFG. The function
`dace.codegen.tools.type_inference.infer_expr_type()` was returning
`int` type for unknown symbols, while we expected `None`. The fix is to
first check if the symbol exists in parent SDFG, in which case we can
directly map it; otherwise, we rely on `infer_expr_type()` to infer the
type.

---------

Co-authored-by: Philip Mueller <philip.mueller@cscs.ch>