bug[next]: fix symlink issue in tmp directory (#2156)
Resolve symlinks to workaround an issue on MacOS where the default tmp
directory is a symlink which might sometimes get resolved in a way we
don't control. The issue showed in the Compiledb, when trying to
substitute CMake generated paths with a relative path.

See #2153 for enabling MacOS in ci.