feat[dace][next]: Switch to DaCe MapFusion (#2148)
Before GT4Py shipped with its own Map fusion transformation, which were
a copy of the one in a [DaCe PR](https://github.com/spcl/dace/pull/1987).
However, the PR has now caught up and the implementation provided inside
it, actually surpasses the transformations that ships with GT4Py.

Thus it was decided to [update the DaCe fork GT4Py is using](https://github.com/GridTools/gt4py/pull/2149)
to include the new version of the transformation and actually using them.

However, GT4Py still creates wrapper around the DaCe versions because,
because some default values needs to be changed and also a callback
mechanism is needed, to restrict the fusion.
However, in the long term, probably a better solution should be found.

In addition the Map fusion transformations are no longer named
serial/parallel but vertical/horizontal, which is also in accordance
with the splitting transformations and DaCe itself.