refactor[dace][next]: Renamed `SerialMapPromoter` to `MapPromoter`. (#2158)
In [PR#2148](https://github.com/GridTools/gt4py/pull/2148) Edaordo
pointed out that with the renaming of `Serial` to `Vertical in Map
fusion, the `SerialMapPromoter` should also be renamed to
`VerticalMapPromoter`. However, it was noticed that the promoter is not
restricted to serial Maps, in fact it was always able to do promotions in
all dimensions, it was just never used until
[PR#2063](https://github.com/GridTools/gt4py/pull/2063).

Thus this PR fixes this issue and renames it to just `MapPromoter` which
is the most accurate description.

Must be merged after
[PR#2148](https://github.com/GridTools/gt4py/pull/2148).