style: update linting tools configuration and apply changes (#2159)
This PR replaces the deprecated `prettier` pre-commit hook by `mdformat`
for the auto-formatting of markdown files, and adds a few more rules to
`ruff check` (some of them in preview mode). All the other changes are a
consequence of running the linting/formatting tools, specially
`mdformat` since it formats the code inside code blocks in markdown
documents.