ci: update development and GitHub Actions CI infrastructure  (#2160)
There are two main changes in this PR related to the development
infrastructure.

First, the dev scripts use now a more standard python package structure
with a `__main__` module, and their dependencies have been moved back to
`pyproject.toml` as a `scripts` dependency group.

Second, change the `test-component` GitHub Actions workflow to be more
explicit about the sessions which are tested, because all test sessions
are listed in the workflow file, and only the (optional) exclusions are
loaded from a dynamically generated JSON file.

Finally, this PR also contains minor documentation cleanups.