feat[next][dace]: Simplify symbolic expression based on domain range (#2174)
This PR adds `target_domain` to `SubgraphContext`, used in the the
lowering to SDFG. This new attribute stores the domain of the target
field found in the top-level `SetAt` expression. The range of the target
field is used to make assumptions on some symbols, such as `range_end >=
range_start`. These assumptions are applied through sympy to simplify
the symbolic shape of temporary arrays and memlet subsets.